v1.2.0 Ahoy ahoy ! Lots of good stuff on this version !
A brand new Firmware upgrade screen, that will allow you to update your watch !

The update of the watch will bring much more stability to the watch in general.

We also redesigned the tutorial to bring lots of important infos and we added a color cue in the notification screen.

Now that the watch's firmware can be upgraded, lots of cool new stuff will come !